+++ meta_package_manager/labels.py
@@ -20,12 +20,10 @@
 from __future__ import annotations
 
 from boltons.iterutils import flatten
-from click_extra.platform import ALL_OS_LABELS
+from click_extra.platform import ANY_PLATFORM, MACOS, ANY_BSD, ANY_LINUX, ANY_UNIX_BUT_MACOS, WINDOWS
 
 from .pool import pool
 
-CLI_NAME = "mpm"
-"""Main CLI ID."""
 
 LABELS = [
     (
@@ -34,59 +32,78 @@ LABELS = [
         "Xbar/SwiftBar plugin code, documentation and features",
     ),
 ]
-"""Structure: ``("label_name", "color", "optional_description")``."""
+"""Global registry of all labels used in the project.
+
+Structure:
+
+.. code-block:: python
+
+    ("label_name", "color", "optional_description")
+"""
 
 
-PLATFORM_LABELS = {}
-""" Maps all platform names to their labels. """
+def generate_labels(all_labels: set[str], groups: dict[str, set[str]], prefix: str, color: str) -> dict[str, str]:
+    """Generate labels."""
+    # Check group definitions.
+    group_entries = tuple(flatten(groups.values()))
+    grouped_labels = set(group_entries)
+    # Check there is no duplucates between groups.
+    assert len(group_entries) == len(grouped_labels)
+    # Check all labels to groups are referenced in the full label set.
+    assert grouped_labels.issubset(all_labels)
+    # Check group IDs do not collide with original labels.
+    assert all_labels.isdisjoint(groups.keys())
+
+    new_labels = {}
+
+    # Create a dedicated label for each non-grouped entry.
+    standalone_labels = all_labels - grouped_labels
+    for label_id in standalone_labels:
+        full_name = f"{prefix}{label_id}"
+        # Check the addition of the prefix does not collide with an existing label.
+        assert full_name not in all_labels
+        new_labels[label_id] = full_name
+        # Register label to the global registry.
+        LABELS.append((full_name, color, label_id))
+
+    # Create a dedicated label for each group.
+    for group_id, label_ids in groups.items():
+        full_name = f"{prefix}{group_id}"
+        # Check the addition of the prefix does not collide with an existing label.
+        assert full_name not in all_labels
+        for label_id in label_ids:
+            new_labels[label_id] = full_name
+        # Register label to the global registry.
+        LABELS.append((full_name, color, ", ".join(sorted(label_ids))))
+
+    return new_labels
 
 
-MANAGER_LABELS = {}
-""" Maps all manager IDs to their labels. """
+# Add mpm as its own manager.
+all_manager_ids = set(pool.all_manager_ids) | {"mpm"}
 
-
-MANAGER_GROUPS = {
+MANAGER_GROUPS: dict[str, set[str]] = {
     "dnf-based": {"dnf", "yum"},
-    "dpkg-based": {"dpkg", "apt", "apt-mint", "opkg"},
+    "dpkg-based": {"apt", "apt-mint", "opkg"},
     "npm-based": {"npm", "yarn"},
     "pacman-based": {"pacman", "pacaur", "paru", "yay"},
     "pip-based": {"pip", "pipx"},
 }
-"""Managers sharing some roots or implementation will be grouped together."""
+"""Managers sharing some origin or implementation are grouped together."""
+
+MANAGER_LABELS = generate_labels(all_manager_ids, MANAGER_GROUPS, "📦 manager: ", "#bfdadc")
+""" Maps all manager IDs to their labels. """
 
 
-PLATFORM_PREFIX = "🖥 platform: "
-"""Default platform label prefix."""
+PLATFORM_GROUPS: dict[str, set[str]] = {
+    # Group all BSD platforms but macOS.
+    "BSD": ANY_BSD - {MACOS},
+    "Linux": ANY_LINUX,
+    "macOS": {MACOS},
+    "Unix": ANY_UNIX_BUT_MACOS - ANY_LINUX - ANY_BSD,
+    "Windows": {WINDOWS},
+}
+"""Similar platforms are grouped together."""
 
-PLATFORM_COLOR = "#bfd4f2"
-"""Default platform label color."""
-
-MANAGER_PREFIX = "📦 manager: "
-"""Default manager label prefix."""
-
-MANAGER_COLOR = "#bfdadc"
-"""Default manager label color."""
-
-
-# Create one label per platform.
-for platform_name in ALL_OS_LABELS:
-    label_id = f"{PLATFORM_PREFIX}{platform_name}"
-    LABELS.append((label_id, PLATFORM_COLOR, platform_name))
-    PLATFORM_LABELS[platform_name] = label_id
-
-
-# Create one label per manager. Add mpm as its own manager.
-non_grouped_managers = set(pool) - set(flatten(MANAGER_GROUPS.values())) | {CLI_NAME}
-for manager_id in non_grouped_managers:
-    label_id = f"{MANAGER_PREFIX}{manager_id}"
-    LABELS.append((label_id, MANAGER_COLOR, manager_id))
-    if manager_id != CLI_NAME:
-        MANAGER_LABELS[manager_id] = label_id
-
-
-# Add labels for grouped managers.
-for group_label, manager_ids in MANAGER_GROUPS.items():
-    label_id = f"{MANAGER_PREFIX}{group_label}"
-    LABELS.append((label_id, MANAGER_COLOR, ", ".join(sorted(manager_ids))))
-    for manager_id in manager_ids:
-        MANAGER_LABELS[manager_id] = label_id
+PLATFORM_LABELS = generate_labels(ANY_PLATFORM, PLATFORM_GROUPS, "🖥 platform: ", "#bfd4f2")
+""" Maps all platform names to their labels. """